requirements.txt
PyMySQL==0.8.1
simplejson==3.15.0
arabic_reshaper==2.0.14
python_bidi==0.4.0
Scrapy==1.5.0
Flask==0.12.2
